refactor memory reduction patches into own folder
