+++ common/src/main/java/net/caffeinemc/mods/lithium/mixin/ai/poi/reduce_poi_memory/package-info.java
@@ -1,0 +1,4 @@
+@MixinConfigOption(description = "Reduce memory consumption of POI system")
+package net.caffeinemc.mods.lithium.mixin.ai.poi.reduce_poi_memory;
+
+import net.caffeinemc.gradle.MixinConfigOption;
