+++ lithium-fabric-mixin-config.md
@@ -41,6 +41,10 @@ Implements a faster POI search
 (default: `true`)  
 Portal search uses the faster POI search and optimized loaded state caching
   
+### `mixin.ai.poi.reduce_poi_memory`
+(default: `true`)  
+Reduce memory consumption of POI system
+  
 ### `mixin.ai.poi.tasks`
 (default: `true`)  
 Mob Tasks which search for POIs use the optimized POI search
